♻️ Refactor code into modular files
