+++ instance-scheduler/utils.go
@@ -1,0 +1,62 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"log"
+	"strings"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
+	"github.com/aws/aws-sdk-go-v2/service/ssm"
+)
+
+type ISSMGetParameter interface {
+	GetParameter(ctx context.Context, params *ssm.GetParameterInput, optFns ...func(*ssm.Options)) (*ssm.GetParameterOutput, error)
+}
+
+func getParameter(client ISSMGetParameter, parameterName string) string {
+	result, err := client.GetParameter(context.TODO(), &ssm.GetParameterInput{
+		Name:           aws.String(parameterName),
+		WithDecryption: aws.Bool(true),
+	})
+	if err != nil {
+		log.Fatal(err)
+	}
+	return *result.Parameter.Value
+}
+
+type ISecretManagerGetSecretValue interface {
+	GetSecretValue(ctx context.Context, params *secretsmanager.GetSecretValueInput, optFns ...func(*secretsmanager.Options)) (*secretsmanager.GetSecretValueOutput, error)
+}
+
+func getSecret(client ISecretManagerGetSecretValue, secretId string) string {
+	result, err := client.GetSecretValue(context.TODO(), &secretsmanager.GetSecretValueInput{
+		SecretId:     aws.String(secretId),
+		VersionStage: aws.String("AWSCURRENT"),
+	})
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	return *result.SecretString
+}
+
+func getNonProductionAccounts(environments string, skipAccountNames string) map[string]string {
+	accounts := make(map[string]string)
+
+	var allAccounts map[string]interface{}
+	json.Unmarshal([]byte(environments), &allAccounts)
+
+	for _, record := range allAccounts {
+		if rec, ok := record.(map[string]interface{}); ok {
+			for key, val := range rec {
+				// Skip if the account's name ends with "-production", for example: performance-hub-production will be skipped
+				if !strings.HasSuffix(key, "-production") && (len(skipAccountNames) < 1 || !strings.Contains(skipAccountNames, key)) {
+					accounts[key] = val.(string)
+				}
+			}
+		}
+	}
+	return accounts
+}
